requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# Graph Database
neo4j>=5.15.0
py2neo>=2021.2.3

# NLP & Entity Linking
spacy>=3.7.0
# Run: python -m spacy download zh_core_web_sm
# Run: python -m spacy download en_core_web_sm

# Database Connectors
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9  # PostgreSQL
pymysql>=1.1.0  # MySQL

# LLM Providers
openai>=1.10.0
anthropic>=0.18.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Logging & Monitoring
loguru>=0.7.0
tenacity>=8.2.0  # Retry logic

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0

# Optional: Vector Store for Hybrid Search
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Web Interface
streamlit>=1.28.0
plotly>=5.18.0